Fixed test. Changed test class name.
